requirements.txt
gcloud==0.17.0
googleapis-common-protos==1.6.0
httplib2==0.14.0
jws==0.1.3
mysql-connector-python==8.0.17
numpy==1.17.3
oauth2client==3.0.0
pandas==0.25.3
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pycryptodome==3.4.3
Pyrebase==3.0.27
python-dateutil==2.8.1
python-jwt==2.0.1
pytz==2019.3
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.13.0
